fix(migration): user_slide_progress.user_id BIGINT UNSIGNED
The 2026_04_29_lesson_slides migration declared user_id as INT UNSIGNED,
but users.id on the production schema is BIGINT UNSIGNED. The FK
constraint failed with errno 150 on a fresh apply — caught when
creating the table on aviatortutor.com production today after the
slide page started 500'ing because the table was never present.

Aligns the column with users.id type so a fresh DB build doesn't
repeat the same FK formation error.

+++ database/migrations/2026_04_29_lesson_slides.sql
@@ -28,7 +28,7 @@ CREATE TABLE IF NOT EXISTS lesson_slides (
 
 CREATE TABLE IF NOT EXISTS user_slide_progress (
     id               BIGINT UNSIGNED PRIMARY KEY AUTO_INCREMENT,
-    user_id          INT UNSIGNED NOT NULL,
+    user_id          BIGINT UNSIGNED NOT NULL,
     lesson_id        BIGINT UNSIGNED NOT NULL,
     slide_id         BIGINT UNSIGNED NOT NULL,
     answered_correct TINYINT(1) DEFAULT 0,
